Rename KnowledgeEntities to KnowledgeDocsClusters
