+++ client/src/views/Graphs/views/Knowledge/KnowledgeDocsClusters.vue
